Move long running test to integration tests
